requirements.txt
absl-py==1.0.0
attrs==21.4.0
certifi==2022.5.18
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.2
mediapipe==0.8.10
numpy==1.22.3
opencv-contrib-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
torch==1.11.0
torchvision==0.12.0
typing-extensions==4.2.0
urllib3==1.26.9
